Separate pl-dircopy process from rest of analysis process
Remove the usage of stagingDcmImages since processing rows can now be determined via plugin status from ChRIS.
In order to have the processing rows show up immediately on the Dashboard after clicking Analyze, without needing to refresh, pl-diropy instances must have been created for each image before routing to the Dashboard.
The changes in this commit split up the image analysis process so that pl-dircopy is first run for all images, then the user is navigated back to the Dashboard, and finally, the rest of the plugins for the analysis is run.