[CNXC-294] Use StudyDate and SeriesDescription on CreateAnalysisPage (#49)
Problem Statement
Currently, the date being displayed under each study is not the study date and the Series are labelled using their DICOM file names. The UI should be reading the StudyDate and SeriesDescription properties from the DICOM metadata and display these instead.

Implementation
With the new updates made to CUBE, getPACSFiles now returns a DICOM metadata object that includes the StudyDate, so this is used to display the appropriate date under each study. The SeriesDescription is now being used instead of the fname to label each series.