Merge pull request #51 from darwinai/feature/CNXC-305-filter-results-by-MRN
When the pl-dircopy instance is created, it is provided the Patient MRN as a title. This allows the UI to filter by MRN by passing the MRN into the name query parameter when calling getFeeds to fetch results to display on the table.
In order to reduce the number of calls made to the backend, lodash debounce was added to debounce the PastAnalysisTable filter state updates.