Correct time comparison to be done holistically
Make comparison happen in milliseconds instead of only checking for days, since the days method
doesn't necessarily represent the total number of days in the duration (eg. if you have only one month,
which is greater than a day, then the original comparison wouldn't work since days would be zero, but
this duration is greater than a day).