Prevent processing visual if running pdf generation
An issue we encounter is the lack of ability to distinguish between
the jobs currently running without making more api requests.
So, a rough process to prevent the processing visual from showing up is
acting like there are no jobs running at all.

Currently there is nothing that handles the case where pdf generation
fails and there is an error (#77 handles only retrieving the model data
but jobsErrored will still be off if pdf generation doesn't work). We
could consider moving the acc.jobsErrored line into the same
conditional, but then we run into the case (which probably can't happen
because you can't access the DICOM viewer in this case) where we can
potentially hide the number of jobs errored if the most recent plugin is
related to pdf generation.